compares/requirements.txt
recbole==1.2.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
pyyaml==6.0.2
wandb==0.19.6
torch==2.5.0 